+++ priv/repo/migrations/20210616095956_add_executable_implementations.exs
@@ -1,0 +1,15 @@
+defmodule TdDd.Repo.Migrations.AddInternalImplementations do
+  use Ecto.Migration
+
+  def up do
+    alter table(:rule_implementations) do
+      add(:executable, :boolean, default: true, null: false)
+    end
+  end
+
+  def down do
+    alter table(:rule_implementations) do
+      remove(:executable, :boolean)
+    end
+  end
+end